backend/requirements/common.txt
django~=4.0.0
django-cors-headers
djangorestframework
pillow


